requirements.txt
pyyaml
requests
faker
htmlbuilder
yattag
Jinja2